Refactored and make code work
