+++ vulnfeeds/cmd/nvd-cve-osv/main.go
@@ -13,6 +13,7 @@ import (
 
 	"golang.org/x/exp/slices"
 
+	"github.com/google/osv/vulnfeeds/common"
 	"github.com/google/osv/vulnfeeds/cves"
 	"github.com/google/osv/vulnfeeds/git"
 	"github.com/google/osv/vulnfeeds/utility"
@@ -104,7 +105,7 @@ var VendorProductDenyList = []VendorProduct{
 // Takes a CVE ID string (for logging), cves.VersionInfo with AffectedVersions and
 // typically no AffectedCommits and attempts to add AffectedCommits (including Fixed commits) where there aren't any.
 // Refuses to add the same commit to AffectedCommits more than once.
-func GitVersionsToCommits(CVE cves.CVEID, versions cves.VersionInfo, repos []string, cache git.RepoTagsCache) (v cves.VersionInfo, e error) {
+func GitVersionsToCommits(CVE cves.CVEID, versions common.VersionInfo, repos []string, cache git.RepoTagsCache) (v common.VersionInfo, e error) {
 	// versions is a VersionInfo with AffectedVersions and typically no AffectedCommits
 	// v is a VersionInfo with AffectedCommits (containing Fixed commits) included
 	v = versions
@@ -118,7 +119,7 @@ func GitVersionsToCommits(CVE cves.CVEID, versions cves.VersionInfo, repos []str
 			Logger.Infof("[%s]: Attempting version resolution for %+v using %q", CVE, av, repo)
 			introducedEquivalentCommit := ""
 			if av.Introduced != "" {
-				ac, err := git.VersionToCommit(av.Introduced, repo, cves.Introduced, normalizedTags)
+				ac, err := git.VersionToCommit(av.Introduced, repo, common.Introduced, normalizedTags)
 				if err != nil {
 					Logger.Warnf("[%s]: Failed to get a Git commit for introduced version %q from %q: %v", CVE, av.Introduced, repo, err)
 				} else {
@@ -134,7 +135,7 @@ func GitVersionsToCommits(CVE cves.CVEID, versions cves.VersionInfo, repos []str
 			if v.HasFixedCommits(repo) && av.Fixed != "" {
 				Logger.Infof("[%s]: Using preassumed fixed commits %+v instead of deriving from fixed version %q", CVE, v.FixedCommits(repo), av.Fixed)
 			} else if av.Fixed != "" {
-				ac, err := git.VersionToCommit(av.Fixed, repo, cves.Fixed, normalizedTags)
+				ac, err := git.VersionToCommit(av.Fixed, repo, common.Fixed, normalizedTags)
 				if err != nil {
 					Logger.Warnf("[%s]: Failed to get a Git commit for fixed version %q from %q: %v", CVE, av.Fixed, repo, err)
 				} else {
@@ -147,7 +148,7 @@ func GitVersionsToCommits(CVE cves.CVEID, versions cves.VersionInfo, repos []str
 			// AffectedCommits (with Fixed commits) when the CVE has appropriate references.
 			lastAffectedEquivalentCommit := ""
 			if !v.HasFixedCommits(repo) && av.LastAffected != "" {
-				ac, err := git.VersionToCommit(av.LastAffected, repo, cves.LastAffected, normalizedTags)
+				ac, err := git.VersionToCommit(av.LastAffected, repo, common.LastAffected, normalizedTags)
 				if err != nil {
 					Logger.Warnf("[%s]: Failed to get a Git commit for last_affected version %q from %q: %v", CVE, av.LastAffected, repo, err)
 				} else {
@@ -159,7 +160,7 @@ func GitVersionsToCommits(CVE cves.CVEID, versions cves.VersionInfo, repos []str
 			// doesn't result in a half-resolved (false positive-causing)
 			// situation with a successfully resolved introduced version and an
 			// unsuccessfully resolved fixed or last_affected version.
-			ac := cves.AffectedCommit{}
+			ac := common.AffectedCommit{}
 			if fixedEquivalentCommit != "" || lastAffectedEquivalentCommit != "" {
 				ac.SetRepo(repo)
 				if introducedEquivalentCommit != "" {
@@ -168,7 +169,7 @@ func GitVersionsToCommits(CVE cves.CVEID, versions cves.VersionInfo, repos []str
 				ac.SetFixed(fixedEquivalentCommit)
 				ac.SetLastAffected(lastAffectedEquivalentCommit)
 			}
-			if ac == (cves.AffectedCommit{}) {
+			if ac == (common.AffectedCommit{}) {
 				// Nothing resolved, move on to the next AffectedVersion
 				Logger.Warnf("[%s]: Sufficient resolution not possible for %+v", CVE, av)
 				continue
@@ -288,7 +289,7 @@ func CVEToOSV(CVE cves.CVE, repos []string, cache git.RepoTagsCache, directory s
 		}
 	}
 
-	slices.SortStableFunc(versions.AffectedCommits, cves.AffectedCommitCompare)
+	slices.SortStableFunc(versions.AffectedCommits, common.AffectedCommitCompare)
 
 	affected := vulns.Affected{}
 	affected.AttachExtractedVersionInfo(versions)
@@ -387,7 +388,7 @@ func CVEToPackageInfo(CVE cves.CVE, repos []string, cache git.RepoTagsCache, dir
 
 	versions.AffectedVersions = nil // these have served their purpose and are not required in the resulting output.
 
-	slices.SortStableFunc(versions.AffectedCommits, cves.AffectedCommitCompare)
+	slices.SortStableFunc(versions.AffectedCommits, common.AffectedCommitCompare)
 
 	var pkgInfos []vulns.PackageInfo
 	pi := vulns.PackageInfo{VersionInfo: versions}
