+++ vulnfeeds/cves/testdata/TestExtractGitCommit_A_GitHub_link_with_tags.yaml
@@ -17,7 +17,7 @@ interactions:
         headers:
             Accept:
                 - text/html
-        url: https://github.com/google/osv.dev/releases/tag/v0.0.14
+        url: https://github.com/redis/redis/releases/tag/6.2.17
         method: HEAD
       response:
         proto: HTTP/2.0
@@ -38,17 +38,17 @@ interactions:
             Content-Type:
                 - text/html; charset=utf-8
             Date:
-                - Thu, 22 May 2025 05:43:02 GMT
+                - Thu, 22 May 2025 05:42:40 GMT
             Etag:
-                - W/"eb81f64035b4867941a1ed13519f619e"
+                - W/"7f1dbc041362432b5e9086ce572a0a9b"
             Referrer-Policy:
                 - no-referrer-when-downgrade
             Server:
                 - github.com
             Set-Cookie:
-                - _gh_sess=rm4IrULBoFcnB%2FwZzF35MwTJJlTTrgaPJ6zDr%2B5stlgZO3blndu%2FxKXU6zXxGTfjf2LLB1r1cvexNRgsZ0If5YnIwy7w%2BC07K4ZpYCW9r3GyDLNohZYwOTjcE8IjSEniKaTYzoIred2yITKotPU3SEELh6SaFwJSFdyEXFkXOjyIfl7q%2FWvH7hFfM9ZSPSM4MO9%2FUo60PQJ34LCbB6gMq8ZQ94IyXj3wMGnA4G%2FH3DuzPqaiBHZU%2BIF0SKnf60Lji5AnvYI8ZIZYxpcSOLYqrw%3D%3D--Cn%2FJz8liWxJh78x%2B--HnC%2BgzxzTwrd2IPxOARF%2Fw%3D%3D; Path=/; HttpOnly; Secure; SameSite=Lax
-                - _octo=GH1.1.1912048022.1747892582; Path=/; Domain=github.com; Expires=Fri, 22 May 2026 05:43:02 GMT; Secure; SameSite=Lax
-                - logged_in=no; Path=/; Domain=github.com; Expires=Fri, 22 May 2026 05:43:02 GMT; HttpOnly; Secure; SameSite=Lax
+                - _gh_sess=hKs29PwNrSztseR0QvKQGrslfW9qKSypwRsZkbOoRQI%2FC%2FVyv%2B%2BcC5cXp2CU1DCpnmFcEjBLsztHiioz%2BsNSMIEdLGjpzmblprJxXBeCQJYs4PQm7FHMEVlfGgN7%2F4MI94qnVdoMleN4PO2jQQZeGtmoYak06QGZLK2nsiKt%2FahqjSIbnjqlpiLs5CGJQYqPMjb%2BUnzE2SCxI76Qtp37pahssSBTBFIicNNCDFM5VBe9wYnMlYTWr%2FhyF0gYjDllIYRg1E3d1DefmEiAjq405A%3D%3D--a8AR4nS%2BGAXMwVEU--XfVHX1uRcBEA2xJcw%2FgNKA%3D%3D; Path=/; HttpOnly; Secure; SameSite=Lax
+                - _octo=GH1.1.1334909775.1747892559; Path=/; Domain=github.com; Expires=Fri, 22 May 2026 05:42:39 GMT; Secure; SameSite=Lax
+                - logged_in=no; Path=/; Domain=github.com; Expires=Fri, 22 May 2026 05:42:39 GMT; HttpOnly; Secure; SameSite=Lax
             Strict-Transport-Security:
                 - max-age=31536000; includeSubdomains; preload
             Vary:
@@ -58,9 +58,9 @@ interactions:
             X-Frame-Options:
                 - deny
             X-Github-Request-Id:
-                - 710E:186D0A:20B1B69:25E65F1:682EB965
+                - 713E:6AC37:20AA583:25DDAC8:682EB94F
             X-Xss-Protection:
                 - "0"
         status: 200 OK
         code: 200
-        duration: 547.394262ms
+        duration: 543.057115ms
